Firing phasers now subtracts from firing Ship's energy. Tests to verify behavior. Resourced out more console text.

+++ StarTrek_KG/Interfaces/ISystem.cs
@@ -4,7 +4,7 @@ namespace StarTrek_KG.Interfaces
 {
     public interface ISystem
     {
-        int Energy { get; set; }
+        double Energy { get; set; }
         bool Destroyed { get; set; }
         Map Map { get; set; }
     }
